Base class for StringSegment - Segment is created. Frequency for DuplicateStringSegmentsWalkerBase is now `long` instead of `int`.

+++ Platform/Platform.Helpers/Collections/SegmentsWalkers/DuplicateStringSegmentsWalkerBase.cs
@@ -11,7 +11,7 @@
         }
 
         protected abstract void OnDublicateFound(ref StringSegment segment);
-        protected abstract int GetSegmentFrequency(ref StringSegment segment);
-        protected abstract void SetSegmentFrequency(ref StringSegment segment, int frequency);
+        protected abstract long GetSegmentFrequency(ref StringSegment segment);
+        protected abstract void SetSegmentFrequency(ref StringSegment segment, long frequency);
     }
 }
